+++ violet-framework/src/main/java/com/horstmann/violet/framework/graphics/content/VerticalLayout.java
@@ -43,8 +43,8 @@ public class VerticalLayout extends Layout
     @Override
     public final void refreshUp()
     {
-        int height = 0;
-        int width = 0;
+        double height = 0;
+        double width = 0;
 
         for (Content content: getContents()) {
             height += content.getHeight();
